+++ lab-1/mappers/deposit_mapper.hpp
@@ -2,8 +2,8 @@
 #define DATABASE_PROGRAMMING_1_DEPOSIT_MAPPER_HPP
 
 #include <vector>
+#include <entities/deposit.hpp>
 #include "sql_executor.hpp"
-#include "entities/deposit.hpp"
 
 class deposit_mapper {
 public:
@@ -20,23 +20,27 @@ public:
                 "    annual decimal      not null\n"
                 ")";
 
-        SQLHSTMT statement = executor_->execute(query);
+        sql_statement statement = executor_->execute(query);
         auto retcode = SQLFetch(statement);
     };
 
-    deposit create(const deposit& deposit_) {
+    deposit* create(deposit* deposit) {
+        if (deposit->id.has_value()) {
+            throw std::runtime_error("[deposit_mapper]: already exists");
+        }
+
         std::wstringstream query_builder;
 
         query_builder <<
                       "insert into deposits(name, term, annual)" << std::endl <<
                       "values" << std::endl << '(' <<
-                      '\'' << deposit_.name << '\'' << ',' <<
-                      deposit_.term << ',' <<
-                      deposit_.annual <<
+                      '\'' << deposit->name << '\'' << ',' <<
+                      deposit->term << ',' <<
+                      deposit->annual <<
                       ')' << std::endl <<
                       "returning id";
 
-        SQLHSTMT statement = executor_->execute(query_builder.str());
+        sql_statement statement = executor_->execute(query_builder.str());
 
         SQLINTEGER id;
 
@@ -44,15 +48,15 @@ public:
 
         retcode = SQLFetch(statement);
 
-        auto new_deposit = deposit_;
-        new_deposit.id = id;
-        return new_deposit;
+        deposit->id = id;
+
+        this->deposits_.push_back(deposit);
+
+        return deposit;
     }
 
-    std::vector<deposit> read() {
-        SQLHSTMT statement = executor_->execute("select * from deposits");
-
-        std::vector<deposit> result;
+    std::vector<deposit*> read() {
+        sql_statement statement = executor_->execute("select * from deposits");
 
         SQLINTEGER id;
         SQLWCHAR name[255];
@@ -68,49 +72,89 @@ public:
             retcode = SQLFetch(statement);
 
             if (retcode == SQL_SUCCESS || retcode == SQL_SUCCESS_WITH_INFO) {
-                result.push_back({ .id = id, .name = name, .term = term, .annual = annual });
+                auto new_deposit = new deposit();
+                new_deposit->id = id;
+                new_deposit->name = name;
+                new_deposit->annual = annual;
+                new_deposit->term = term;
+
+                bool found = false;
+                for (auto deposit: deposits_) {
+                    if (deposit->id.value() == new_deposit->id.value()) {
+                        deposit->id = id;
+                        deposit->name = name;
+                        deposit->annual = annual;
+                        deposit->term = term;
+                        found = true;
+                    }
+                }
+
+                if (!found) {
+                    deposits_.push_back(new_deposit);
+                } else {
+                    delete new_deposit;
+                }
             }
             else if (retcode == SQL_NO_DATA) {
                 break;
             }
         }
 
-        return result;
+        return deposits_;
     }
 
-    deposit read(int id_) {
+    deposit* read(const int& position) {
+        if (position > deposits_.size()) {
+            throw std::runtime_error("[deposit_mapper]: out of bounds");
+        }
+
+        return deposits_[position];
+    }
+
+    deposit* read_id(const int& id) {
+        deposit* result = nullptr;
+
+        // Ищем локально
+        for (auto deposit: deposits_) {
+            if (deposit->id == id) {
+                result = deposit;
+                return result;
+            }
+        }
+
+        // Если не нашли - получаем из БД
         std::stringstream query_builder;
 
         query_builder <<
                       "select * from deposits" << std::endl <<
-                      "where id = " << id_;
+                      "where id = " << id;
 
-        SQLHSTMT statement = executor_->execute(query_builder.str());
+        sql_statement statement = executor_->execute(query_builder.str());
 
-        SQLINTEGER id;
         SQLWCHAR name[255];
         SQLINTEGER term;
         SQLREAL annual;
 
-        auto retcode = SQLBindCol(statement, 1, SQL_C_LONG, &id, 0, nullptr);
-        retcode = SQLBindCol(statement, 2, SQL_C_WCHAR, &name, 255, nullptr);
+        auto retcode = SQLBindCol(statement, 2, SQL_C_WCHAR, &name, 255, nullptr);
         retcode = SQLBindCol(statement, 3, SQL_C_LONG, &term, 0, nullptr);
         retcode = SQLBindCol(statement, 4, SQL_C_FLOAT, &annual, 0, nullptr);
 
         retcode = SQLFetch(statement);
 
-        deposit fetched = {
-                .id = id,
-                .name = name,
-                .term = term,
-                .annual = annual,
-        };
+        // Сохранили локально
+        result = new deposit();
+        result->id = id;
+        result->name = name;
+        result->annual = annual;
+        result->term = term;
 
-        return fetched;
+        deposits_.push_back(result);
+
+        return result;
     }
 
-    deposit update(const deposit& deposit_) {
-        if (!deposit_.id.has_value()) {
+    deposit* update(deposit* deposit) {
+        if (!deposit->id.has_value()) {
             throw std::runtime_error("[deposit_mapper]: record doesn't exists");
         }
 
@@ -119,13 +163,13 @@ public:
         query_builder <<
                       "update deposits" << std::endl <<
                       "set " <<
-                      "name=" << '\'' << deposit_.name << '\'' << ',' <<
-                      "term=" << deposit_.term << ',' <<
-                      "annual=" << deposit_.annual << std::endl <<
-                      "where id = " << deposit_.id.value() << std::endl <<
+                      "name=" << '\'' << deposit->name << '\'' << ',' <<
+                      "term=" << deposit->term << ',' <<
+                      "annual=" << deposit->annual << std::endl <<
+                      "where id = " << deposit->id.value() << std::endl <<
                       "returning *";
 
-        SQLHSTMT statement = executor_->execute(query_builder.str());
+        sql_statement statement = executor_->execute(query_builder.str());
 
         SQLINTEGER id;
         SQLWCHAR name[255];
@@ -139,50 +183,40 @@ public:
 
         retcode = SQLFetch(statement);
 
-        deposit fetched = {
-                .id = id,
-                .name = name,
-                .term = term,
-                .annual = annual
-        };
+        deposit->name = name;
+        deposit->term = term;
+        deposit->annual = annual;
 
-        return fetched;
+        return deposit;
     }
 
-    deposit remove(int id_) {
+    int remove(const int& position) {
+        if (position > deposits_.size()) {
+            throw std::runtime_error("[deposit_mapper]: out of bounds");
+        }
+
         std::stringstream query_builder;
 
         query_builder <<
                       "delete from deposits" << std::endl <<
-                      "where id = " << id_ << std::endl <<
-                      "returning *";
+                      "where id = " << deposits_[position]->id.value() << std::endl <<
+                      "returning id";
 
-        SQLHSTMT statement = executor_->execute(query_builder.str());
+        sql_statement statement = executor_->execute(query_builder.str());
 
         SQLINTEGER id;
-        SQLWCHAR name[255];
-        SQLINTEGER term;
-        SQLREAL annual;
-
         auto retcode = SQLBindCol(statement, 1, SQL_C_LONG, &id, 0, nullptr);
-        retcode = SQLBindCol(statement, 2, SQL_C_WCHAR, &name, 255, nullptr);
-        retcode = SQLBindCol(statement, 3, SQL_C_LONG, &term, 0, nullptr);
-        retcode = SQLBindCol(statement, 4, SQL_C_FLOAT, &annual, 0, nullptr);
-
         retcode = SQLFetch(statement);
 
-        deposit fetched = {
-                .id = id,
-                .name = name,
-                .term = term,
-                .annual = annual
-        };
+        delete deposits_[position];
+        deposits_.erase(deposits_.begin() + position);
 
-        return fetched;
+        return id;
     }
 
 private:
     sql_executor* executor_;
+    std::vector<deposit*> deposits_;
 };
 
 #endif //DATABASE_PROGRAMMING_1_DEPOSIT_MAPPER_HPP
